F76 - Add Inventory Screen (#64)
This commit brought in a lot of UI layout changes that will be built upon for the completion of F91. This code works pretty well and resizes well on desktop, although the code feels wrong to me for whatever reason. Maybe I'll find a better system for building UI and solid resizing using js or better cs. For now, this system works, and we're gonna roll with it.